+++ app/app_pub_test.go
@@ -100,7 +100,7 @@ func setupAppTest(t *testing.T) (*assert.Assertions, *require.Assertions, *Binan
 	pub.IsLive = true
 
 	keeper := app.DexKeeper
-	keeper.CollectOrderInfoForPublish = true
+	keeper.EnablePublish()
 	tradingPair := dextypes.NewTradingPair("XYZ-000", "BNB", 102000)
 	keeper.PairMapper.AddTradingPair(app.DeliverState.Ctx, tradingPair)
 	keeper.AddEngine(tradingPair)
@@ -114,8 +114,8 @@ func setupAppTest(t *testing.T) (*assert.Assertions, *require.Assertions, *Binan
 	keeper.FeeManager.FeeConfig.CancelFee = 12
 	keeper.FeeManager.FeeConfig.CancelFeeNative = 6
 
-	_, buyerAcc := testutils.NewAccountForPub(ctx, am, 100000000000, 0, 0) // give user enough coins to pay the fee
-	_, sellerAcc := testutils.NewAccountForPub(ctx, am, 100000000000, 0, 0)
+	_, buyerAcc := testutils.NewAccountForPub(ctx, am, 100000000000, 0, 0, "XYZ-000") // give user enough coins to pay the fee
+	_, sellerAcc := testutils.NewAccountForPub(ctx, am, 100000000000, 0, 0, "XYZ-000")
 	return assert.New(t), require.New(t), app, buyerAcc, sellerAcc
 }
 
@@ -142,7 +142,7 @@ func TestAppPub_MatchOrder(t *testing.T) {
 
 	ctx := app.DeliverState.Ctx
 	msg := orderPkg.NewNewOrderMsg(buyerAcc.GetAddress(), orderPkg.GenerateOrderID(1, buyerAcc.GetAddress()), orderPkg.Side.BUY, "XYZ-000_BNB", 102000, 300000000)
-	handler := orderPkg.NewHandler(app.GetCodec(), app.DexKeeper, app.AccountKeeper)
+	handler := orderPkg.NewHandler(app.DexKeeper)
 	app.DeliverState.Ctx = app.DeliverState.Ctx.WithBlockHeight(41).WithBlockTime(time.Unix(0, 100))
 	buyerAcc.SetSequence(1)
 	app.AccountKeeper.SetAccount(ctx, buyerAcc)
@@ -216,7 +216,7 @@ func TestAppPub_MatchOrder(t *testing.T) {
 
 func TestAppPub_MatchAndCancelFee(t *testing.T) {
 	assert, require, app, buyerAcc, sellerAcc := setupAppTest(t)
-	handler := orderPkg.NewHandler(app.GetCodec(), app.DexKeeper, app.AccountKeeper)
+	handler := orderPkg.NewHandler(app.DexKeeper)
 	ctx := app.DeliverState.Ctx
 
 	// ==== Place a to-be-matched sell order and a to-be-cancelled buy order (in different symbol)