+++ plugins/dex/client/cli/list.go
@@ -32,13 +32,13 @@ func listTradingPairCmd(cdc *wire.Codec) *cobra.Command {
 			}
 
 			baseAsset := viper.GetString(flagBaseAsset)
-			err = types.ValidateMapperTokenSymbol(baseAsset)
+			err = types.ValidateTokenSymbol(baseAsset)
 			if err != nil {
 				return err
 			}
 
 			quoteAsset := viper.GetString(flagQuoteAsset)
-			err = types.ValidateMapperTokenSymbol(quoteAsset)
+			err = types.ValidateTokenSymbol(quoteAsset)
 			if err != nil {
 				return err
 			}
@@ -71,6 +71,55 @@ func listTradingPairCmd(cdc *wire.Codec) *cobra.Command {
 	cmd.Flags().String(flagQuoteAsset, "", "symbol of the quote currency")
 	cmd.Flags().String(flagInitPrice, "", "init price for this pair")
 	cmd.Flags().Int64(flagProposalId, 0, "list proposal id")
+
+	return cmd
+}
+
+func listMiniTradingPairCmd(cdc *wire.Codec) *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "list-mini",
+		Short: "",
+		RunE: func(cmd *cobra.Command, args []string) error {
+			cliCtx, txbldr := client.PrepareCtx(cdc)
+
+			from, err := cliCtx.GetFromAddress()
+			if err != nil {
+				return err
+			}
+
+			baseAsset := viper.GetString(flagBaseAsset)
+			err = types.ValidateMiniTokenSymbol(baseAsset)
+			if err != nil {
+				return err
+			}
+
+			quoteAsset := viper.GetString(flagQuoteAsset)
+			if quoteAsset != types.NativeTokenSymbol && !strings.HasPrefix(quoteAsset, "BUSD") {
+				return errors.New("invalid quote asset")
+			}
+
+			baseAsset = strings.ToUpper(baseAsset)
+			quoteAsset = strings.ToUpper(quoteAsset)
+
+			initPriceStr := viper.GetString(flagInitPrice)
+			initPrice, err := utils.ParsePrice(initPriceStr)
+			if err != nil {
+				return err
+			}
+
+			msg := list.NewListMiniMsg(from, baseAsset, quoteAsset, initPrice)
+			err = client.SendOrPrintTx(cliCtx, txbldr, msg)
+			if err != nil {
+				return err
+			}
+
+			return nil
+		},
+	}
+
+	cmd.Flags().StringP(flagBaseAsset, "s", "", "symbol of the base asset")
+	cmd.Flags().String(flagQuoteAsset, "", "symbol of the quote currency")
+	cmd.Flags().String(flagInitPrice, "", "init price for this pair")
 
 	return cmd
 }