+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## 0.7.0
+FEATURES
+* [\#725](https://github.com/binance-chain/node/pull/725) [Token] [Dex] BEP8 - Mini-BEP2 token features
+* [\#710](https://github.com/binance-chain/node/pull/710) [DEX] BEP70 - Support busd pair listing and trading
+
+IMPROVEMENTS
+* [\#704](https://github.com/binance-chain/node/pull/704) [DEX] BEP67 Price-based Order Expiration
+* [\#714](https://github.com/binance-chain/node/pull/714) [DEX] Add pendingMatch flag to orderbook query response
+
+## 0.6.3-hf.1
+
+BUG FIXES
+* [\#693](https://github.com/binance-chain/node/pull/693) [Deps] hot fix for hard fork in stdTx getSigner
+
 ## 0.6.3
 
 BUG FIXES